refactor: factor NewEmptyTestDB out of NewTestDB
Factor the creation of an empty test DB out into a separate function and
move it into pkg/db.

This makes it available to tests in pkg/db without the cyclic dependency
on migrations.
